requirements.txt
PyYAML>=6.0,<7.0
pytest>=7.4
pytest-cov>=4.1
